+++ Swanki/Model/Card.swift
@@ -50,10 +50,21 @@ public final class Card {
   /// - learning: integer timestamp
   public var due: Date?
   public var interval: TimeInterval = 0
-  public var factor: Double = 0
+  public var factor: Double = 2.5
   public var reps = 0
   public var lapses = 0
   public var left = 0
+
+  public static func allCards(deck: Deck? = nil) -> Predicate<Card> {
+    if let deck {
+      let deckID = deck.id
+      return #Predicate { card in
+        card.deck?.id == deckID
+      }
+    } else {
+      return #Predicate { _ in true }
+    }
+  }
 
   public static func newCards(deck: Deck? = nil) -> Predicate<Card> {
     if let deck {