+++ Swanki/View/NoteList.swift
@@ -24,9 +24,18 @@ struct NoteList: View {
 
   var body: some View {
     @Bindable var applicationNavigation = applicationNavigation
-    Table(notes, selection: $applicationNavigation.selectedNote) {
-      TableColumn("Spanish", value: \.front.defaultEmpty)
-      TableColumn("English", value: \.back.defaultEmpty)
+    let stats = summaryStatistics
+    VStack {
+      DeckProgressChart(new: stats.newCardCount, learning: stats.learningCardCount, mastered: stats.masteredCardCount)
+        .frame(height: 50)
+        .padding()
+      if atRiskXP > 0 {
+        Text("You are at risk of losing **\(atRiskXP) XP**")
+      }
+      Table(notes, selection: $applicationNavigation.selectedNote) {
+        TableColumn("Spanish", value: \.front.defaultEmpty)
+        TableColumn("English", value: \.back.defaultEmpty)
+      }
     }
     .navigationTitle(deck.name)
     .inspector(isPresented: $isShowingInspector, content: {
@@ -57,6 +66,34 @@ struct NoteList: View {
         .keyboardShortcut(.delete)
         .disabled(applicationNavigation.selectedNote == nil)
       }
+    }
+  }
+
+  private var summaryStatistics: SummaryStatistics {
+    guard let modelContext = deck.modelContext else {
+      logger.warning("Trying to get statistics when the deck has not been saved")
+      return SummaryStatistics()
+    }
+    do {
+      return try modelContext.summaryStatistics(deck: deck)
+    } catch {
+      logger.error("Error fetching summary statistics: \(error)")
+      return SummaryStatistics()
+    }
+  }
+
+  private var atRiskXP: Int {
+    guard let modelContext = deck.modelContext else {
+      logger.warning("Trying to get XP when the deck has not been saved")
+      return 0
+    }
+    do {
+      let todayStats = try modelContext.summaryStatistics(deck: deck)
+      let tomorrowStats = try modelContext.summaryStatistics(on: .now.addingTimeInterval(.day), deck: deck)
+      return todayStats.xp - tomorrowStats.xp
+    } catch {
+      logger.error("Error fetching summary statistics: \(error)")
+      return 0
     }
   }
 
@@ -102,6 +139,5 @@ private struct SelectDeckView: View {
   NavigationStack {
     SelectDeckView()
   }
-  .environment(ApplicationNavigation.previews)
-  .modelContainer(.previews)
+  .withPreviewEnvironment()
 }